+++ [qbx]/[JOB]/jim-mechanic/configs/repairbench_conf.lua
@@ -62,6 +62,7 @@ Config.Emergency = {
         ["Speakers"] = true,
         ["WindowTints"] = true,
         ["Rims"] = true,
+        ["Lightbar"] = true,
     },
 
     PreformaceTable = {