+++ [qbx]/[JOB]/jim-mechanic/server/commands.lua
@@ -10,10 +10,26 @@ onResourceStart(function()
 	registerCommand("cleancar", { locale("serverFunctions", "cleanVehicle"), {}, false, function(source) TriggerClientEvent(getScript()..":client:cleanVehicle", source, true) end})
 	registerCommand("hood", { locale("serverFunctions", "toggleHood"), {}, false, function(source) TriggerClientEvent(getScript()..":client:openDoor", source, nil, 4) end })
 	registerCommand("trunk", { locale("serverFunctions", "toggleTrunk"), {}, false, function(source) TriggerClientEvent(getScript()..":client:openDoor", source, nil, 5) end })
-	registerCommand("door", { locale("serverFunctions", "toggleDoor"), {{name="1-4", help="Door ID"}}, false, function(source, args) TriggerClientEvent(getScript()..":client:openDoor", source, args[1]) end })
+	registerCommand("door", {
+		locale("serverFunctions", "toggleDoor"),
+		{{name="1-4", help="Door ID"}},
+		true,
+		function(source, args)
+			local arg = args["1-4"] or args[1]
+			TriggerClientEvent(getScript()..":client:openDoor", source, arg)
+		end
+	})
 
 	if Config.Harness.HarnessControl then
-		registerCommand("seat", { locale("serverFunctions", "changeSeat"), {{name="id", help="Seat ID"}}, false, function(source, args) TriggerClientEvent(getScript()..":client:seat", source, args[1]) end })
+		registerCommand("seat", {
+			locale("serverFunctions", "changeSeat"),
+			{{name="id", help="Seat ID"}},
+			true,
+			function(source, args)
+				local arg = args["id"] or args[1]
+				TriggerClientEvent(getScript()..":client:seat", source, arg)
+			end
+		})
 	end
 
 	if isStarted(OXLibExport) then  -- if ox_lib allow previewRGB